+++ mapsforge/src/org/mapsforge/preprocessing/map/osmosis/SutherlandHodgmanClipping.java
@@ -42,13 +42,13 @@ class SutherlandHodgmanClipping {
 
 			if (isStartPointInside) {
 				if (clippedPolyline.isEmpty()) {
-					clippedPolyline.add(x1);
-					clippedPolyline.add(y1);
+					clippedPolyline.add(Float.valueOf(x1));
+					clippedPolyline.add(Float.valueOf(y1));
 				}
 
 				if (isEndPointInside) {
-					clippedPolyline.add(x2);
-					clippedPolyline.add(y2);
+					clippedPolyline.add(Float.valueOf(x2));
+					clippedPolyline.add(Float.valueOf(y2));
 				} else {
 					float[] intersection = computeIntersection(edge, x1, y1, x2, y2);
 					clippedPolyline.add(Float.valueOf(intersection[0]));
@@ -58,8 +58,8 @@ class SutherlandHodgmanClipping {
 				float[] intersection = computeIntersection(edge, x1, y1, x2, y2);
 				clippedPolyline.add(Float.valueOf(intersection[0]));
 				clippedPolyline.add(Float.valueOf(intersection[1]));
-				clippedPolyline.add(x2);
-				clippedPolyline.add(y2);
+				clippedPolyline.add(Float.valueOf(x2));
+				clippedPolyline.add(Float.valueOf(y2));
 			}
 		}
 
@@ -69,17 +69,15 @@ class SutherlandHodgmanClipping {
 
 		float[] retVal = new float[clippedPolyline.size()];
 		for (int i = 0; i < clippedPolyline.size(); ++i) {
-			retVal[i] = clippedPolyline.get(i);
+			retVal[i] = clippedPolyline.get(i).floatValue();
 		}
 		return retVal;
 	}
 
-	private static float[] computeIntersection(int[] edge, double x1, double y1, double x2,
-			double y2) {
+	private static float[] computeIntersection(int[] edge, double x1, double y1, double x2, double y2) {
 		if (edge[1] == edge[3]) {
 			// horizontal edge
-			return new float[] { (float) (x1 + (edge[1] - y1) * ((x2 - x1) / (y2 - y1))),
-					edge[1] };
+			return new float[] { (float) (x1 + (edge[1] - y1) * ((x2 - x1) / (y2 - y1))), edge[1] };
 		}
 		// vertical edge
 		return new float[] { edge[0], (float) (y1 + (edge[0] - x1) * ((y2 - y1) / (x2 - x1))) };
@@ -114,16 +112,16 @@ class SutherlandHodgmanClipping {
 	 */
 	static float[] clipPolyline(float[] polyline, int[] rectangle) {
 		// bottom edge
-		float[] clippedPolyline = clipPolylineToEdge(polyline, new int[] { rectangle[0],
-				rectangle[1], rectangle[2], rectangle[1] });
+		float[] clippedPolyline = clipPolylineToEdge(polyline, new int[] { rectangle[0], rectangle[1],
+				rectangle[2], rectangle[1] });
 		// right edge
-		clippedPolyline = clipPolylineToEdge(clippedPolyline, new int[] { rectangle[2],
-				rectangle[1], rectangle[2], rectangle[3] });
+		clippedPolyline = clipPolylineToEdge(clippedPolyline, new int[] { rectangle[2], rectangle[1],
+				rectangle[2], rectangle[3] });
 		// top edge
-		clippedPolyline = clipPolylineToEdge(clippedPolyline, new int[] { rectangle[2],
-				rectangle[3], rectangle[0], rectangle[3] });
+		clippedPolyline = clipPolylineToEdge(clippedPolyline, new int[] { rectangle[2], rectangle[3],
+				rectangle[0], rectangle[3] });
 		// left edge
-		return clipPolylineToEdge(clippedPolyline, new int[] { rectangle[0], rectangle[3],
-				rectangle[0], rectangle[1] });
+		return clipPolylineToEdge(clippedPolyline, new int[] { rectangle[0], rectangle[3], rectangle[0],
+				rectangle[1] });
 	}
 }
